requirements.txt
Django==5.2.6
mysql-connector==2.2.9
virtualenv
dj-database-url==0.5.0
gunicorn==20.0.4
whitenoise==5.2.0
Pillow==11.3.0
requests==2.32.5
django-cors-headers==4.8.0
numpy==2.2.6
scikit-learn==1.7.1
google-genai==1.46.0
